LPS-89474 Add segments to SoyContext only for Pages
