+++ modules/apps/layout/layout-content-page-editor-web/src/main/java/com/liferay/layout/content/page/editor/web/internal/display/context/ContentPageLayoutEditorDisplayContext.java
@@ -14,7 +14,13 @@
 
 package com.liferay.layout.content.page.editor.web.internal.display.context;
 
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.template.soy.util.SoyContext;
+import com.liferay.portal.template.soy.util.SoyContextFactoryUtil;
+import com.liferay.segments.model.SegmentsEntry;
+import com.liferay.segments.service.SegmentsEntryServiceUtil;
+
+import java.util.List;
 
 import javax.portlet.RenderResponse;
 
@@ -41,6 +47,8 @@ public class ContentPageLayoutEditorDisplayContext
 
 		SoyContext soyContext = super.getEditorContext();
 
+		soyContext.put(
+			"availableSegments", _getSoyContextAvailableSegmentsEntries());
 		soyContext.put("sidebarPanels", getSidebarPanelSoyContexts(false));
 
 		_editorSoyContext = soyContext;
@@ -48,6 +56,52 @@ public class ContentPageLayoutEditorDisplayContext
 		return _editorSoyContext;
 	}
 
+	@Override
+	public SoyContext getFragmentsEditorToolbarContext()
+		throws PortalException {
+
+		if (_fragmentsEditorToolbarSoyContext != null) {
+			return _fragmentsEditorToolbarSoyContext;
+		}
+
+		SoyContext soyContext = super.getFragmentsEditorToolbarContext();
+
+		soyContext.put(
+			"availableSegments", _getSoyContextAvailableSegmentsEntries());
+
+		_fragmentsEditorToolbarSoyContext = soyContext;
+
+		return _fragmentsEditorToolbarSoyContext;
+	}
+
+	private SoyContext _getSoyContextAvailableSegmentsEntries()
+		throws PortalException {
+
+		SoyContext availableSegmentsEntriesSoyContext =
+			SoyContextFactoryUtil.createSoyContext();
+
+		List<SegmentsEntry> segmentsEntries =
+			SegmentsEntryServiceUtil.getSegmentsEntries(getGroupId());
+
+		for (SegmentsEntry segmentsEntry : segmentsEntries) {
+			SoyContext segmentsSoyContext =
+				SoyContextFactoryUtil.createSoyContext();
+
+			segmentsSoyContext.put(
+				"segmentLabel",
+				segmentsEntry.getName(themeDisplay.getLocale()));
+			segmentsSoyContext.put(
+				"segmentId", segmentsEntry.getSegmentsEntryId());
+			segmentsSoyContext.put("segmentKey", segmentsEntry.getKey());
+
+			availableSegmentsEntriesSoyContext.put(
+				segmentsEntry.getKey(), segmentsSoyContext);
+		}
+
+		return availableSegmentsEntriesSoyContext;
+	}
+
 	private SoyContext _editorSoyContext;
+	private SoyContext _fragmentsEditorToolbarSoyContext;
 
 }
